dashboard/requirements.txt
# Dashboard Web - Dependencias del entorno envTFM
dash==3.2.0
dash-bootstrap-components==2.0.3
dash-leaflet==1.1.3
plotly==6.2.0

# Análisis de datos
pandas==2.3.1
numpy==2.2.6

# APIs y web requests
requests==2.32.4

# Geoespacial (necesario para KML y mapas)
geopandas==1.1.1
shapely==2.1.1
pyproj==3.7.1

# Archivos y utilidades
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2

# Imágenes
pillow==11.3.0

# HTTP server
flask==3.1.1
werkzeug==3.1.3

# Google API (para integración con Telegram bot)
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.178.0

# Utilidades básicas
click==8.2.1
packaging==25.0